Upgraded google-cloud-firestore to the latest version (0.61.0-beta) and added a FirebaseOptions#setFireStoreTimestampsInSnapshotsEnabled(boolean) method.
The FirestoreOptions#setTimestampsInSnapshotsEnabled(boolean) method is then called in the FirestoreClient constructor
with the value passed to the FirebaseOptions#setFireStoreTimestampsInSnapshotsEnabled(boolean) method.

This allows for firebase-admin-java users to adapt to the Timestamps being returned in DocumentSnapshots as added in
https://github.com/GoogleCloudPlatform/google-cloud-java/pull/3317

com.google.api:gax was added as a dependency to the pom.xml file as there were
"java.lang.NoClassDefFoundError: com/google/api/gax/rpc/ClientStream" errors when running "mvn test"
without this dependency.

A couple of tests were also added to the FirestoreClientTest class.