Fix CreateRequest chaining and provider ID extraction. (#399)
This makes the base `CreateRequest` setters return the proper instance type, so that methods can be chained. This also makes it so that the provider ID can be parsed from the resource name.

A package private `getProviderId()` method was also added, which will be needed by the `FirebaseUserManager` class when the provider config operations are added there.

Also renamed `AuthProviderConfig` to `ProviderConfig` since "Auth" is redundant with the package name.

This work is part of adding multi-tenancy support (see issue #332).